Update README
Adding the image, link and description of the project on this commit to the README file. I also just saw part of the class from Thursday that I had to miss where Nikki stated we need to do more, smaller commits as opposed to one large commit. I will make this a priority going forward. Thank you!